@@ -9,9 +9,9 @@
     }
   ],
   "require": {
-    "php": ">=5.4",
-    "doctrine/dbal": "~2.0",
-    "psr/log": "~1.0",
+    "php": ">=8.0",
+    "doctrine/dbal": "~3.0",
+    "psr/log": "~3.0",
     "ext-json": "*",
     "ext-zlib": "*",
     "ext-curl": "*"
